% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_add_method.R
\name{add,PopulationChangeParameters,PopulationChangeParameters-method}
\alias{add,PopulationChangeParameters,PopulationChangeParameters-method}
\title{Add Two PopulationChangeParameters Objects Together}
\usage{
\S4method{add}{PopulationChangeParameters,PopulationChangeParameters}(obj1, obj2)
}
\arguments{
\item{obj1}{\code{PopulationChangeParameters} object}

\item{obj2}{\code{PopulationChangeParameters} object}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Add Two PopulationChangeParameters Objects Together
}
