% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_scenarios_manage.R
\name{AddScenario}
\alias{AddScenario}
\title{Add A New Scenario To A Scenarios Table}
\usage{
AddScenario(tbl, UniqueID, ...)
}
\arguments{
\item{tbl}{Scenarios table}

\item{UniqueID}{Scenario name. Must be unique.}

\item{...}{Field values (set Details)}
}
\value{
Updated scenarios table
}
\description{
Add A New Scenario To A Scenarios Table
}
\details{
Allowed fields (in the ... arguments):
\itemize{
\item \strong{WeeksPerYr} (double)
\item \strong{HrsPerWeek} (double)
\item \strong{BaselinePop} (double)
\item \strong{PopType} (character)
\item \strong{o_PopGrowth} (logical)
\item \strong{o_Seasonality} (logical)
\item \strong{o_Fertility_decr} (logical)
\item \strong{o_MHIVTB_decr} (logical)
\item \strong{o_ChildDis_decr} (logical)
\item \strong{sheet_TaskValues} (character)
\item \strong{sheet_PopValues} (character)
\item \strong{sheet_SeasonalityCurves} (character)
\item \strong{sheet_sheet_Cadre} (character)
\item \strong{DeliveryModel} (character)
\item \strong{Geography_dontedit} (character)
}
}
\examples{
\dontrun{
tbl <- ehep::CreateScenariosTable()
tbl <- ehep::AddScenario(tbl, "Scenario 1")
}
}
