% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_time.R
\name{ClinicalTaskTime}
\alias{ClinicalTaskTime}
\title{Calculate Annual Time For Clinical Task}
\usage{
ClinicalTaskTime(tasks, taskID, demographics, year)
}
\arguments{
\item{tasks}{Dataframe of task parameters (as returned by \code{loadTaskParameters})}

\item{taskID}{Task ID string}

\item{demographics}{List of population pyramid dataframes}

\item{year}{Year (index into \code{demographics} list)}
}
\value{
Annual time in minutes
}
\description{
Calculate Annual Time For Clinical Task
}
