% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_time.R
\name{ClinicalTaskTimesGroup}
\alias{ClinicalTaskTimesGroup}
\title{Calculate Clinical Task Times}
\usage{
ClinicalTaskTimesGroup(tasks, taskIDs, demographics, years)
}
\arguments{
\item{tasks}{Dataframe of task parameters (as returned by \code{loadTaskParameters})}

\item{taskIDs}{Vector of task ID strings}

\item{demographics}{List of population pyramid dataframes}

\item{years}{Vector of years (usually \code{globalPackageEnvironment$years})}
}
\value{
Dataframe of annual times in minutes
}
\description{
Calculate clinical task times for a group of tasks over a spread of years
}
