% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_predict.R
\name{computeBirths}
\alias{computeBirths}
\title{Compute Births}
\usage{
computeBirths(female_population, rates)
}
\arguments{
\item{female_population}{Vector of female population pyramid}

\item{rates}{Exploded vector of annual fertility rates}
}
\value{
Total number of expected births
}
\description{
Compute the total number of births for a year, given the female population
pyramid and the annual fertility rates per age.
}
