% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_post_process.R
\name{ComputeCadreAllocations}
\alias{ComputeCadreAllocations}
\title{Compute Cadre Allocations}
\usage{
ComputeCadreAllocations(DR = NULL)
}
\arguments{
\item{DR}{Task time results from \code{ReadAndCollateSuiteResults()}}
}
\value{
data.table
}
\description{
Different healthcare delivery models involve different mixtures of types
of healthcare workers ("Cadres"). \code{ComputeCadresAllocations()} reads
cadre information from the input Excel file, then combines the allocation
percentages with computed task times to produce a table showing how much
time each cadre spends per year on each healthcare task.
}
\examples{
\dontrun{
#' resultsFiles <- c("results/results_BasicModel_Jul04.csv",
                  "results/results_ComprehensiveModel_Jul04.csv",
                  "results/results_MergedModel_Jul04.csv")

DR <- ehep::ReadAndCollateSuiteResults(files = resultsFiles)
CA <- ehep:::ComputeCadreAllocations(DR)
}
}
