% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_predict.R
\name{computeDeaths}
\alias{computeDeaths}
\title{Compute Deaths}
\usage{
computeDeaths(population, rates)
}
\arguments{
\item{population}{Dataframe of female and male population pyramids, as
returned by, for example, \code{loadInitialPopulation}}

\item{rates}{List with exploded vectors of annual death rates, as returned by
\code{explodeMortalityRates}}
}
\value{
List of expected deaths, one vector for each sex
}
\description{
Compute the total number of deaths for a year, given the population
pyramids for females and males, and the annual rates per age for both
sexes.
}
