% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_predict.R
\name{ComputeDemographicsProjection}
\alias{ComputeDemographicsProjection}
\title{Compute Demographics Projection}
\usage{
ComputeDemographicsProjection(
  initial_population_pyramid,
  fertility_rates,
  mortality_rates,
  years,
  debug = FALSE
)
}
\arguments{
\item{fertility_rates}{Fertility rates}

\item{mortality_rates}{Mortality rates}

\item{years}{Vector of years to model}

\item{debug}{Flag for debugging output}
}
\value{
Demographics time-series
}
\description{
Use an initial population pyramid, fertility rates, and mortality rates
to predict future population pyramids.
}
