% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_post_process.R
\name{ComputeSummaryStats}
\alias{ComputeSummaryStats}
\title{Compute Summary Statistics}
\usage{
ComputeSummaryStats(DR = NULL, CA = NULL)
}
\arguments{
\item{DR}{From \code{ReadAndCollateSuiteResults()} function}

\item{CA}{From \code{ComputeCadreAllocations()} function}
}
\value{
List of computed summary tables
}
\description{
Compute Summary Statistics
}
