% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_scenarios_manage.R
\name{DeleteScenario}
\alias{DeleteScenario}
\title{Delete A Scenario From A Scenarios Table}
\usage{
DeleteScenario(tbl, UniqueID)
}
\arguments{
\item{tbl}{Scenarios table}

\item{UniqueID}{Scenario name. Must be unique.}
}
\value{
Updated scenarios table
}
\description{
Delete A Scenario From A Scenarios Table
}
