% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot_fertility.R
\name{.combineFRatesMatrices}
\alias{.combineFRatesMatrices}
\title{Extract And Combine Fertility Matrices Across Multiple Trials}
\usage{
.combineFRatesMatrices(results)
}
\arguments{
\item{results}{Results list (as returned by \code{RunExperiments()})}
}
\value{
Data frame of fertility rates across all trials
}
\description{
Extract And Combine Fertility Matrices Across Multiple Trials
}
