% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_predict.R
\name{.computeDemographicsProjection}
\alias{.computeDemographicsProjection}
\title{Compute Demographics Projection}
\usage{
.computeDemographicsProjection(
  initial_population_pyramid,
  fertility_rates,
  mortality_rates,
  years
)
}
\arguments{
\item{initial_population_pyramid}{Population pyramids dataframe. Must have \code{$Male}
and \code{$Female} fields}

\item{fertility_rates}{Fertility rates}

\item{mortality_rates}{Mortality rates}

\item{years}{Vector of years to model}
}
\value{
Demographics time-series
}
\description{
Use an initial population pyramid, fertility rates, and mortality rates
to predict future population pyramids
}
