% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot_fertility.R
\name{.fRatesBoxPlot}
\alias{.fRatesBoxPlot}
\title{Plot Fertility Rates Statistics As Box Plot}
\usage{
.fRatesBoxPlot(df)
}
\arguments{
\item{results}{Results list (as returned by \code{RunExperiments()})}
}
\value{
ggplot grob
}
\description{
Plot Fertility Rates Statistics As Box Plot
}
