% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot_fertility.R
\name{.fRatesRibbonPlot}
\alias{.fRatesRibbonPlot}
\title{Plot Fertility Rates Statistics As Ribbon Plot}
\usage{
.fRatesRibbonPlot(df, se, trials)
}
\arguments{
\item{se}{Whether to show standard error intervals}

\item{trials}{Number of trials}

\item{results}{Results list (as returned by \code{RunExperiments()})}
}
\value{
ggplot grob
}
\description{
Plot Fertility Rates Statistics As Ribbon Plot
}
