% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_predict.R
\name{explodeFertilityRates}
\alias{explodeFertilityRates}
\title{Convert Fertility Rates From Banded To Per-Age}
\usage{
explodeFertilityRates(bandedAnnualRates)
}
\arguments{
\item{bandedAnnualRates}{Rates reported in age buckets}
}
\value{
List of vectors of per-year-of-age rates, for males and females
}
\description{
Birth rates are reported in age bands - 15-19 years, 20-29 years, etc.
\code{explodeFertilityRates} converts the vector of banded rates into a
vector with one rate per year of age.
}
