% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_predict.R
\name{explodeMortalityRates}
\alias{explodeMortalityRates}
\title{Convert Mortality Rates From Banded To Per-Age}
\usage{
explodeMortalityRates(bandedAnnualRates)
}
\arguments{
\item{bandedAnnualRates}{Rates reported in age buckets}
}
\value{
List of vectors of per-year-of-age rates, for males and females
}
\description{
Mortality rates are reported in age bands - 1-4 years, 5-9 years, etc.
\code{explodeMortalityRates} converts the vector of banded rates into a
vector with one rate per year of age.
}
