% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{gatherPopulation}
\alias{gatherPopulation}
\title{Convert Population Pyramids Into One Dataframe}
\usage{
gatherPopulation(pops)
}
\arguments{
\item{pops}{List of population pyramids for several years}
}
\value{
Long skinny dataframe of population information
}
\description{
Rearrange the demographics structure from a list of dataframes with combined
male/female population data into one long, skinny dataframe with Year and
Gender fields.
}
