% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_input.R
\name{generateFertilityRates}
\alias{generateFertilityRates}
\title{Generate Time-Series of Fertility Rates}
\usage{
generateFertilityRates(popChangeParamsList = NULL)
}
\arguments{
\item{sheetName}{Population parameters list (see \code{loadPopulationChangeParameters}).}
}
\value{
Dataframe of annual fertility rates
}
\description{
Take the initial value and change rate information returned be a call to
\code{loadPopulationChangeParameters} and derive a database of annual
fertility rates stratified by population age cohort.
}
