% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_frates_matrix.R
\name{generateFertilityRatesMatrix}
\alias{generateFertilityRatesMatrix}
\title{Generate A Matrix Of Stochastically Varied Fertility Rates}
\usage{
generateFertilityRatesMatrix()
}
\value{
Rates matrix
}
\description{
Generate A Matrix Of Stochastically Varied Fertility Rates
}
