% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_config.R
\name{generateMortalityRates}
\alias{generateMortalityRates}
\title{Generate Time-Series of Mortality Rates}
\usage{
generateMortalityRates(popChangeParamsList = NULL)
}
\arguments{
\item{sheetName}{Population parameters list (see \code{loadPopulationChangeParameters}).}
}
\value{
Dataframe of annual mortality rates
}
\description{
Take the initial value and change rate information returned by a call to
\code{loadPopulationChangeParameters} and derive a database of annual
mortality rates stratified by age.
}
