% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_mrates_matrix.R
\name{generateMortalityRatesMatrix}
\alias{generateMortalityRatesMatrix}
\title{Generate A Matrix Of Stochastically Varied Mortality Rates}
\usage{
generateMortalityRatesMatrix()
}
\value{
Rates matrix
}
\description{
Generate A Matrix Of Stochastically Varied Mortality Rates
}
