% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_getFertilityRates_method.R
\name{getFertilityRates,PopulationChangeParameters-method}
\alias{getFertilityRates,PopulationChangeParameters-method}
\title{Get Fertility Rates From A PopulationChangeParameters Object}
\usage{
\S4method{getFertilityRates}{PopulationChangeParameters}(object)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}
}
\value{
Labeled vector of fertility rates
}
\description{
Get Fertility Rates From A PopulationChangeParameters Object
}
