% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_getMortalityRates_method.R
\name{getMortalityRates,PopulationChangeParameters-method}
\alias{getMortalityRates,PopulationChangeParameters-method}
\title{Get Mortality Rates From A PopulationChangeParameters Object}
\usage{
\S4method{getMortalityRates}{PopulationChangeParameters}(object)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}
}
\value{
Labeled vector of mortality rates
}
\description{
Get Mortality Rates From A PopulationChangeParameters Object
}
