% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_config.R
\name{InitializeHealthcareTasks}
\alias{InitializeHealthcareTasks}
\title{Initialize Healthcare Task Information}
\usage{
InitializeHealthcareTasks(...)
}
\arguments{
\item{...}{Parameters passed through to \code{\link[=loadTaskParameters]{loadTaskParameters()}}}
}
\value{
NULL (invisible)
}
\description{
Read the healthcare task information from the model inputs Excel file, and
save to a location in the global package environment.
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
