% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_scenarios_config.R
\name{InitializeScenarios}
\alias{InitializeScenarios}
\title{Initialize Experiment Scenario Information}
\usage{
InitializeScenarios(...)
}
\arguments{
\item{...}{See \code{loadScenarios()}}
}
\value{
NULL (invisible)
}
\description{
Read the healthcare task information from the model inputs Excel file, and
save to a location in the global package environment.
}
