% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_seasonality_config.R
\name{InitializeSeasonality}
\alias{InitializeSeasonality}
\title{Initialize Seasonality Information}
\usage{
InitializeSeasonality(...)
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
NULL (invisible)
}
\description{
Read the healthcare seasonality information from the model inputs Excel file, and
save to a location in the global package environment.
}
