% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_input.R
\name{loadInitialPopulation}
\alias{loadInitialPopulation}
\title{Load Initial Population}
\usage{
loadInitialPopulation(sheetName = "TotalPop")
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
List with three \code{PopulationPyramid} objects:
\code{female}, \code{male} and \code{total}
}
\description{
Read the initial population pyramid from the model inputs Excel file. The
name and location of the model inputs Excel file is loaded from the global
configuration JSON file.
}
