% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_input.R
\name{loadPopulationChangeParameters}
\alias{loadPopulationChangeParameters}
\title{Load Population Change Parameters}
\usage{
loadPopulationChangeParameters(sheetName = "PopValues")
}
\arguments{
\item{sheetName}{Sheet name from model input Excel file.}
}
\value{
List with two \code{PopulationChangeParameters} objects:
\code{initValues} and \code{changeRates}
}
\description{
Load Population Change Parameters
}
