% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotFertilityRates}
\alias{PlotFertilityRates}
\title{Plot Fertility Rates}
\usage{
PlotFertilityRates(populationRates, year)
}
\arguments{
\item{populationRates}{Population rates list, as returned by \code{RunExperiments()}}

\item{year}{Year}
}
\value{
ggplot grob
}
\description{
Plot Fertility Rates
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)

g <- ehep::PlotFertilityRates(results[[49]]$PopulationRates, 2030)
print(g)
}
}
