% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotMortalityRates}
\alias{PlotMortalityRates}
\title{Plot Mortality Rates}
\usage{
PlotMortalityRates(ratesMatrix, year)
}
\arguments{
\item{ratesMatrix}{Matrix of banded mortality rates versus years}

\item{year}{Year}
}
\value{
ggplot grob
}
\description{
Plot Mortality Rates
}
