% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotPopulationCurve}
\alias{PlotPopulationCurve}
\title{Plot A Single Population Curve}
\usage{
PlotPopulationCurve(pop, xaxis = NULL, color = .colorM, title = "")
}
\arguments{
\item{pop}{Population values as a numeric vector}

\item{xaxis}{Vector of age values for the x axis}

\item{color}{Plot color}

\item{title}{Plot title}
}
\value{
ggplot grob
}
\description{
Plot A Single Population Curve
}
