% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotPopulationCurves}
\alias{PlotPopulationCurves}
\title{Plot A Family Of Population Curves}
\usage{
PlotPopulationCurves(..., xaxis = NULL, colors = NULL, shapes = NULL)
}
\arguments{
\item{...}{Population vectors}

\item{xaxis}{Vector of age values for the x axis}

\item{colors}{Vector of colors for each population line}

\item{shapes}{Vector of shapes for each population point}
}
\value{
ggplot grob
}
\description{
Plot A Family Of Population Curves
}
