% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotResultsPopulationCurve}
\alias{PlotResultsPopulationCurve}
\title{Plot A Single Population Curve From A Results List}
\usage{
PlotResultsPopulationCurve(results, trial = 1, year = 2020, sex = "f", ...)
}
\arguments{
\item{results}{Results list (as returned by \code{RunExperiments()})}

\item{trial}{Trail number (index into the results list)}

\item{year}{Year in trial timeseries to plot}

\item{sex}{Sex to plot (from c("m", "M", "f", "F"), defaults to female)}

\item{...}{Parameters passed to \code{\link[=PlotPopulationCurve]{PlotPopulationCurve()}}}
}
\value{
ggplot grob, or NULL on error
}
\description{
Plot A Single Population Curve From A Results List
}
