% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_post_process.R
\name{ReadAndCollateSuiteResults}
\alias{ReadAndCollateSuiteResults}
\title{Prepare Suite Results For Post-Processing}
\usage{
ReadAndCollateSuiteResults(
  files = NULL,
  dir = NULL,
  pattern = "^.*\\\\.csv",
  configFile = NULL,
  preProcFunc = .defPreProcFunc,
  postProcFunc = .defPostProcFunc
)
}
\arguments{
\item{files}{Vector of results files to read. (default = NULL)}

\item{dir}{Directory of results files to read. The \strong{files} parameter has
precedence over the \strong{dir} parameter. (default = NULL)}

\item{pattern}{Regular expression to select which results files will be read
from directory \strong{dir}. (default = "^.*\\.csv", simple filter for CSV files)}

\item{configFile}{Input configuration Excel file. If no \strong{configFile} is
specified, the one declared during global initialization of the package
is used. (default = NULL)}

\item{preProcFunc}{(default = .defPreProcFunc)}

\item{postProcFunc}{(default = .defPostProcFunc)}
}
\value{
Data table.
}
\description{
Combine a family of suite results into one big file, and join with scenario
and task value data for further data analysis.
}
