% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_scenarios_manage.R
\name{ReadScenario}
\alias{ReadScenario}
\title{Read A Scenario Record}
\usage{
ReadScenario(tbl, UniqueID)
}
\arguments{
\item{tbl}{Scenarios table}

\item{UniqueID}{Scenario name}
}
\value{
Scenario record (as a tibble)
}
\description{
Read A Scenario Record
}
