% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_experiment.R
\name{RunExperiment}
\alias{RunExperiment}
\title{Run An EHEP Modeling Experiment}
\usage{
RunExperiment(scenarioName = "ScenarioA", debug = FALSE)
}
\arguments{
\item{scenarioName}{(default = "ScenarioA")}

\item{debug}{(default = FALSE)}
}
\value{
List of dataframes of per-task times, or NULL
}
\description{
Combine the parameter values in the Base and Epsilon environments, save the
new parameter values to \code{experimentValuesEnvironment}, then run a full
set of calculations.
}
\details{
Results are written back into \code{experimentValuesEnvironment}.
}
