% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_experiments.R
\name{RunExperiments}
\alias{RunExperiments}
\title{Run A Suite Of EHEP Modeling Experiments}
\usage{
RunExperiments(scenarioName = "ScenarioA", trials = 100, debug = FALSE)
}
\arguments{
\item{scenarioName}{(default = "ScenarioA")}

\item{trials}{(default = 100)}

\item{debug}{(default = FALSE)}
}
\value{
List of dataframes of per-task times, or NULL
}
\description{
Results are written back into \code{experimentValuesEnvironment}.
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
