% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_exp_control.R
\name{SaveBaseSettings}
\alias{SaveBaseSettings}
\title{Save Base Values For A Stochastic Experiment}
\usage{
SaveBaseSettings(scenarioName = "")
}
\arguments{
\item{scenarioName}{Scenario name string}
}
\value{
Scenario information, or NULL if the scenario is invalid
}
\description{
Copy configuration parameters from the Global Package Environment to the
Base Values Environment.
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

result <- ehep::SaveBaseSettings(scenario)
}
}
