% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_save_results.R
\name{SaveResults}
\alias{SaveResults}
\title{Save Experiment Results As CSV File}
\usage{
SaveResults(results, filepath, scenario, trial, run)
}
\arguments{
\item{results}{TBD}

\item{filepath}{TBD}

\item{scenario}{TBD}

\item{trial}{TBD}

\item{run}{TBD}
}
\value{
NULL (invisible)
}
\description{
Save Experiment Results As CSV File
}
