% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_save_results.R
\name{SaveSuiteDemographics}
\alias{SaveSuiteDemographics}
\title{Save Demographics Information From A Suite Of Experiments}
\usage{
SaveSuiteDemographics(results, filename = "out.csv", breaks = NULL)
}
\arguments{
\item{results}{Results structure}

\item{filename}{CSV file to write to}

\item{breaks}{Vector of population bucket boundaries}
}
\value{

}
\description{
Save Demographics Information From A Suite Of Experiments
}
