% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_save_results.R
\name{SaveSuiteResults}
\alias{SaveSuiteResults}
\title{Save Experiment Results As CSV File}
\usage{
SaveSuiteResults(results, filepath, scenario, run)
}
\arguments{
\item{results}{Results structure as returned by \code{RunExperiments()}}

\item{filepath}{Location to write CSV file}

\item{scenario}{Name of experiments scenario, as passed to \code{RunExperiments()}}

\item{run}{Run ID}
}
\value{
NULL (invisible)
}
\description{
Save Experiment Results As CSV File
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)

runId <- 1

ehep::SaveSuiteResults(results, "results.csv", scenario, runId)
}
}
