% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_setFertilityRates_method.R
\name{setFertilityRates,PopulationChangeParameters,numeric-method}
\alias{setFertilityRates,PopulationChangeParameters,numeric-method}
\title{Set Fertility Rates On A PopulationChangeParameters Object}
\usage{
\S4method{setFertilityRates}{PopulationChangeParameters,numeric}(object, values)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}

\item{values}{}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Set Fertility Rates On A PopulationChangeParameters Object
}
