% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_pop_change_params_class.R
\name{setFromVector}
\alias{setFromVector}
\title{Set Population Change Parameters}
\usage{
setFromVector(object, values)
}
\arguments{
\item{object}{PopulationChangeParameters object}

\item{values}{Vector of values to copy into PopulationChangeParameters object}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Set Population Change Parameters
}
