% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_config.R
\name{SetGlobalStartEndYears}
\alias{SetGlobalStartEndYears}
\title{Set Global Start And End Year Parameters}
\usage{
SetGlobalStartEndYears(start = 2020, end = 2040)
}
\arguments{
\item{start}{Starting year}

\item{end}{Ending year (must be greater than \code{start} )}
}
\value{
Nothing
}
\description{
Set Global Start And End Year Parameters
}
