% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_setMortalityRates_method.R
\name{setMortalityRates,PopulationChangeParameters,numeric-method}
\alias{setMortalityRates,PopulationChangeParameters,numeric-method}
\title{Set Mortality Rates On A PopulationChangeParameters Object}
\usage{
\S4method{setMortalityRates}{PopulationChangeParameters,numeric}(object, values)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}

\item{values}{}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Set Mortality Rates On A PopulationChangeParameters Object
}
