% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_pop_change_params_class.R
\name{setValue}
\alias{setValue}
\title{Set Population Change Parameter Values}
\usage{
setValue(object, label, value)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}

\item{label}{Name of variable to write}

\item{value}{Value to write to variable}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Set Population Change Parameter Values
}
