% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_params_class.R
\docType{class}
\name{TaskParameters-class}
\alias{TaskParameters-class}
\alias{TaskParameters}
\title{Task Parameters Class}
\value{
Class of type \code{TaskParameters}
}
\description{
Task Parameters Class
}
\section{Slots}{

\describe{
\item{\code{values}}{numeric}

\item{\code{labels}}{character}
}}

