% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_time.R
\name{TaskTime}
\alias{TaskTime}
\title{Calculate Annual Time For A Task}
\usage{
TaskTime(taskID, year, debug = FALSE)
}
\arguments{
\item{taskID}{Index of row in task tables}

\item{year}{Year (index into \code{demographics} list)}

\item{debug}{Emit debugging information (default = FALSE)}
}
\value{
Annual time in minutes
}
\description{
Calculate Annual Time For A Task
}
