% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_time.R
\name{TaskTimesGroup}
\alias{TaskTimesGroup}
\title{Calculate Task Times For A Group Of Tasks}
\usage{
TaskTimesGroup(taskIDs, years)
}
\arguments{
\item{taskIDs}{Vector of task indices into \code{globalPackageEnvironment$taskData}}

\item{years}{Vector of years (usually \code{globalPackageEnvironment$years})}
}
\value{
List with two matrices:
Time - annual task times in minutes
N - number of times task was performed
}
\description{
Calculate clinical task times for a group of tasks over a spread of years
}
