% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_scenarios_manage.R
\name{UpdateScenario}
\alias{UpdateScenario}
\title{Update A Scenario Record}
\usage{
UpdateScenario(tbl, UniqueID, ...)
}
\arguments{
\item{tbl}{Scenarios table}

\item{UniqueID}{Scenario name}

\item{...}{Field values (set Details)}
}
\value{
Updated scenarios table
}
\description{
Update A Scenario Record
}
\details{
Allowed fields (in the ... arguments):
\itemize{
\item \strong{WeeksPerYr} (double)
\item \strong{HrsPerWeek} (double)
\item \strong{BaselinePop} (double)
\item \strong{PopType} (character)
\item \strong{o_PopGrowth} (logical)
\item \strong{o_Seasonality} (logical)
\item \strong{o_Fertility_decr} (logical)
\item \strong{o_MHIVTB_decr} (logical)
\item \strong{o_ChildDis_decr} (logical)
}
}
