% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_zero_method.R
\name{zero,PopulationChangeParameters-method}
\alias{zero,PopulationChangeParameters-method}
\title{Zero A PopulationChangeParameters Object}
\usage{
\S4method{zero}{PopulationChangeParameters}(object)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Zero A PopulationChangeParameters Object
}
