% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_pop_change_params_class.R
\name{zero}
\alias{zero}
\title{Zero A PopulationChangeParameters Object}
\usage{
zero(object)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Zero A PopulationChangeParameters Object
}
