% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_lint.R
\name{CheckInputExcelFileFormat}
\alias{CheckInputExcelFileFormat}
\title{Perform Sanity Checks On Input Excel File}
\usage{
CheckInputExcelFileFormat(
  inputFile = NULL,
  outputFile = NULL,
  scenarioSheet = "Scenarios",
  seasonalityOffsetsSheet = "SeasonalityOffsets"
)
}
\arguments{
\item{inputFile}{Excel file to examine. If NULL, check the file defined in the global Configuration.}

\item{outputFile}{Results output file. If NULL, results are printed to the console.}

\item{scenariosSheet}{Name of the tab/sheet with scenario details. Default = "Scenarios".}
}
\value{
Error code.
0 = Success
< 0 (negative) = Fatal error. Nothing will work right.
> 0 (positive) = Warning. Some scenarios might fail, some might succeed.
}
\description{
Do a set of sanity checks on the input Excel file.
}
