% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_time.R
\name{ClinicalTaskTime}
\alias{ClinicalTaskTime}
\title{Calculate Annual Time For Clinical Task}
\usage{
ClinicalTaskTime(taskID, year)
}
\arguments{
\item{taskID}{Index of row in task tables}

\item{year}{Year (index into \code{demographics} list)}
}
\value{
Annual time in minutes
}
\description{
Calculate Annual Time For Clinical Task
}
