% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_time.R
\name{ClinicalTaskTimesGroup}
\alias{ClinicalTaskTimesGroup}
\title{Calculate Clinical Task Times}
\usage{
ClinicalTaskTimesGroup(taskIDs, years)
}
\arguments{
\item{taskIDs}{Vector of task indices into \code{globalPackageEnvironment$taskData}}

\item{years}{Vector of years (usually \code{globalPackageEnvironment$years})}
}
\value{
Dataframe of annual times in minutes
}
\description{
Calculate clinical task times for a group of tasks over a spread of years
}
