% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_predict.R
\name{ComputeDemographicsProjection}
\alias{ComputeDemographicsProjection}
\title{Compute Demographics Projection}
\usage{
ComputeDemographicsProjection(
  initial_population_pyramid,
  fertility_rates,
  mortality_rates,
  years,
  normalize = NULL,
  growthFlag = TRUE,
  debug = FALSE
)
}
\arguments{
\item{initial_population_pyramid}{Population pyramids dataframe. Must have
\code{$Age}, \code{$Male} and \code{$Female} fields.}

\item{fertility_rates}{Fertility rates matrix}

\item{mortality_rates}{Mortality rates matrix}

\item{years}{Vector of years to model}

\item{normalize}{Whether or not to normalize the initial population (default = NULL)}

\item{growthFlag}{If FALSE, normalize each year to the same population as
the initial year (default = TRUE)}

\item{debug}{Flag for debugging output}
}
\value{
Demographics time-series
}
\description{
Use an initial population pyramid, fertility rates, and mortality rates
to predict future population pyramids.
}
