% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_predict.R
\name{ComputeDemographicsProjection}
\alias{ComputeDemographicsProjection}
\title{Compute Demographics Projection}
\usage{
ComputeDemographicsProjection(...)
}
\value{
Demographics time-series
}
\description{
Use the population pyramid, fertility rates, and mortality rates loaded
globally to predict future population pyramids. Drops through to
\code{.computeDemographicsProjection()}
}
