% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_exp_control.R
\name{ConfigureExperimentValues}
\alias{ConfigureExperimentValues}
\title{Combine Base and Epsilon Values}
\usage{
ConfigureExperimentValues()
}
\description{
Create the configuration values for an individual model experiment by
adding together values from the Base Values Environment and the Epsilon
Values Environment.
}
