% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_task_config.R
\name{InitializeHealthcareTasks}
\alias{InitializeHealthcareTasks}
\title{Initialize Healthcare Task Information}
\usage{
InitializeHealthcareTasks(...)
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
NULL (invisible)
}
\description{
Read the healthcare task information from the model inputs Excel file, and
save to a location in the global package environment.
}
