% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_stochasticity_config.R
\name{InitializeStochasticParameters}
\alias{InitializeStochasticParameters}
\title{Initialize Stochastic Parameters}
\usage{
InitializeStochasticParameters(...)
}
\arguments{
\item{...}{Parameters passed through to \code{\link[=loadStochasticParameters]{loadStochasticParameters()}}}
}
\value{
NULL (invisible)
}
\description{
Read the stochasticity information from the model inputs Excel file, and
save to a location in the global package environment.
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
