% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_stochasticity_config.R
\name{loadStochasticParameters}
\alias{loadStochasticParameters}
\title{Load Stochastic Parameters}
\usage{
loadStochasticParameters(sheetName = "StochasticParamaters")
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
List with three \code{PopulationPyramid} objects:
\code{female}, \code{male} and \code{total}
}
\description{
Load Stochastic Parameters
}
