% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotFertilityRates}
\alias{PlotFertilityRates}
\title{Plot Fertility Rates}
\usage{
PlotFertilityRates(ratesMatrix, year)
}
\arguments{
\item{ratesMatrix}{Matrix of banded fertility rates versus years}

\item{year}{Year}
}
\value{
ggplot grob
}
\description{
Plot Fertility Rates
}
