% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotFertilityRatesStats}
\alias{PlotFertilityRatesStats}
\title{Plot Fertility Rates Statistics}
\usage{
PlotFertilityRatesStats(results, se = FALSE)
}
\arguments{
\item{results}{Results list (as returned by \code{RunExperiments()})}

\item{se}{Whether to show standard error intervals, or projection intervals}
}
\value{
ggplot grob
}
\description{
Plot Fertility Rates Statistics
}
