% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotPyramid}
\alias{PlotPyramid}
\title{Plot A Single Population Pyramid}
\usage{
PlotPyramid(df, year)
}
\arguments{
\item{df}{Long, skinny dataframe of population data}

\item{year}{Year to graph}
}
\value{
Graphics grob
}
\description{
Plot A Single Population Pyramid
}
