% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot.R
\name{PlotResultsMortalityRates}
\alias{PlotResultsMortalityRates}
\title{Plot A Single Pair of Mortality Rates Curves From A Results List}
\usage{
PlotResultsMortalityRates(results, trial = 1, year = 2020)
}
\arguments{
\item{results}{Results list (as returned by \code{RunExperiments()})}

\item{trial}{Trail number (index into the results list)}

\item{year}{Year in trial timeseries to plot}
}
\value{
ggplot grob, or NULL on error
}
\description{
Plot A Single Pair of Mortality Rates Curves From A Results List
}
