% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_experiment.R
\name{RunExperiment}
\alias{RunExperiment}
\title{Run An EHEP Modeling Experiment}
\usage{
RunExperiment(scenarioName = "ScenarioA", normalize = NULL)
}
\arguments{
\item{scenarioName}{(default = "ScenarioA")}

\item{normalize}{Whether or not to normalize the initial population (default = NULL)}
}
\value{
NULL (invisible)
}
\description{
Combine the parameter values in the Base and Epsilon environments, save the
new parameter values to \code{experimentValuesEnvironment}, then run a full
set of calculations.
}
\details{
Results are written back into \code{experimentValuesEnvironment}.
}
