% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_save_results.R
\name{SaveResults}
\alias{SaveResults}
\title{Save Experiment Results As CSV File}
\usage{
SaveResults(results, filepath = "out.csv", scenario, trial, run)
}
\arguments{
\item{results}{Results structure (as returned by [RunExperiments()])}

\item{filepath}{CSV file to write to (default = "out.csv")}

\item{scenario}{TBD}

\item{trial}{TBD}

\item{run}{TBD}
}
\value{
NULL (invisible)
}
\description{
Save Experiment Results As CSV File
}
