% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_save_results.R
\name{SaveSuiteDemographics}
\alias{SaveSuiteDemographics}
\title{Save Demographics Information From A Suite Of Experiments}
\usage{
SaveSuiteDemographics(results, filepath = "out.csv", breaks = NULL)
}
\arguments{
\item{results}{Results structure (as returned by \code{\link[=RunExperiments]{RunExperiments()}})}

\item{filepath}{CSV file to write to (default = "out.csv")}

\item{breaks}{Vector of population bucket boundaries}
}
\value{
Generated output as a data frame. AgeBucket values can be used to
summarize the output, e.g. with dplyr.
}
\description{
Save Demographics Information From A Suite Of Experiments
}
\examples{
\dontrun{
scenario <- "ScenarioName"

results <-
 ehep::RunExperiments(scenarioName = scenario,
                      trials = 100)

df <- ehep::SaveSuiteDemographics(results, breaks = c(50))

df <- df \%>\%
  dplyr::group_by(Trial, Year, AgeBucket) \%>\%
  dplyr::summarize(Female = sum(Female), Male = sum(Male))
}
}
