% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_save_results.R
\name{SaveSuiteResults}
\alias{SaveSuiteResults}
\title{Save Experiment Results As CSV File}
\usage{
SaveSuiteResults(results, filepath, scenario, run)
}
\arguments{
\item{results}{TBD}

\item{filepath}{TBD}

\item{scenario}{TBD}

\item{run}{TBD}
}
\value{
NULL (invisible)
}
\description{
Save Experiment Results As CSV File
}
